finish OpenID + AnonymousClient
