+++ pac4j-openid/src/test/java/org/pac4j/openid/run/RunYahooOpenIdClient.java
@@ -31,7 +31,7 @@ public class RunYahooOpenIdClient  extends RunClient {
 
     @Override
     protected String getPassword() {
-        return "testpwdscribeup";
+        return "testpwdpac4j";
     }
 
     @Override
